Allow variables to be double instead of int

+++ expression_engine/expression_engine.h
@@ -34,7 +34,7 @@ bool is_left_unary(double (*un_op)(double));
 void create_tree(ListNode **list_head);
 void create_linked_list(ListNode **list_head, int length);
 void convert_tokens_to_nodes(Node **head, char (*tokens)[TOKEN_AMOUNT][TOKEN_LENGTH], int tokens_amount);
-void substitute_variable(Node *node, char var, int val);
+void substitute_variable(Node *node, char var, double val);
 double evaluate_tree(Node *node);
 void delete_tree(Node *node);
